gnu parallel run

+++ run.py
@@ -1,5 +1,6 @@
 # Import libraries for convolutional neural network and data processing
 # We will use pytorch to create a convolutional neural network training on MNIST dataset.
+import os
 import random
 import numpy as np
 import pandas as pd
@@ -7,6 +8,7 @@ import matplotlib.pyplot as plt
 import seaborn as sns
 import torch
 import pickle
+import argparse
 
 # Import the MNIST dataset
 from torchvision import datasets, transforms
@@ -15,12 +17,28 @@ import torch.nn as nn
 from torch.autograd import Variable
 from basic_CNN import CNN
 
-device = torch.device("cuda:2" if torch.cuda.is_available() else "cpu")
-
+# device = torch.device("cuda:2" if torch.cuda.is_available() else "cpu")
+dataset = "mnist"
+seed = 10
+random.seed(seed)
+torch.manual_seed(seed)
 # CIFAR 10 stats
 mean = [0.4914, 0.4822, 0.4465]
 std = [0.2470, 0.2435, 0.2616]
 
+
+parser = argparse.ArgumentParser(description="Fit different DCAs.")
+parser.add_argument("--cuda", type=int, help="save directory")
+parser.add_argument("--experiment", type=int, default=0, help="Experiment number")
+args = parser.parse_args()
+args.cuda = ((args.cuda % 2) + 6)
+device = args.cuda
+# args.experiment=26
+
+def prepare_configs_loop(*all_param_grids):
+    params = np.meshgrid(*all_param_grids, indexing="ij")
+    params = [param.flatten() for param in params]
+    return params
 
 class AddGaussianNoise(object):
     def __init__(self, mean=0.0, std=1.0):
@@ -69,67 +87,77 @@ class AddMask(object):
         return self.__class__.__name__ + "(mask_size={0})".format(self.mask_size)
 
 
-def get_dataloaders():
-    train_data = datasets.MNIST(
-        root="data", train=True, download=True, transform=transforms.ToTensor()
-    )
-    train_data = datasets.CIFAR10(
-        root="data",
-        train=True,
-        download=True,
-        transform=transforms.Compose(
-            [transforms.ToTensor(), transforms.Normalize(mean, std)]
-        ),
-    )
-    test_data = datasets.MNIST(
-        root="data", train=False, download=True, transform=transforms.ToTensor()
-    )
-    test_data = datasets.CIFAR10(
-        root="data",
-        train=False,
-        download=True,
-        transform=transforms.Compose(
-            [transforms.ToTensor(), transforms.Normalize(mean, std)]
-        ),
-    )
-    test_data_noisy = datasets.MNIST(
-        root="data",
-        train=False,
-        download=True,
-        transform=transforms.Compose(
-            [
-                transforms.ToTensor(),
-                transforms.Normalize((0.1307,), (0.3081,)),
-                AddGaussianNoise(0.0, 1.0),
-            ]
-        ),
-    )
-    test_data_noisy = datasets.CIFAR10(
-        root="data",
-        train=False,
-        download=True,
-        transform=transforms.Compose(
-            [
-                transforms.ToTensor(),
-                transforms.Normalize(mean, std),
-                AddGaussianNoise(0.0, 1.0),
-            ]
-        ),
-    )
-    test_data_masked = datasets.MNIST(
-        root="data",
-        train=False,
-        download=True,
-        transform=transforms.Compose([transforms.ToTensor(), AddMask(0.5, True)]),
-    )
-    test_data_masked = datasets.CIFAR10(
-        root="data",
-        train=False,
-        download=True,
-        transform=transforms.Compose(
-            [transforms.ToTensor(), transforms.Normalize(mean, std), AddMask(0.5, True)]
-        ),
-    )
+def get_dataloaders(dataset):
+    if dataset == "mnist":
+        train_data = datasets.MNIST(
+            root="data", train=True, download=True, transform=transforms.ToTensor()
+        )
+
+        test_data = datasets.MNIST(
+            root="data", train=False, download=True, transform=transforms.ToTensor()
+        )
+        test_data_noisy = datasets.MNIST(
+            root="data",
+            train=False,
+            download=True,
+            transform=transforms.Compose(
+                [
+                    transforms.ToTensor(),
+                    transforms.Normalize((0.1307,), (0.3081,)),
+                    AddGaussianNoise(0.0, 1.0),
+                ]
+            ),
+        )
+        test_data_masked = datasets.MNIST(
+            root="data",
+            train=False,
+            download=True,
+            transform=transforms.Compose([transforms.ToTensor(), AddMask(0.5, True)]),
+        )
+
+    elif dataset == "cifar10":
+        train_data = datasets.CIFAR10(
+            root="data",
+            train=True,
+            download=True,
+            transform=transforms.Compose(
+                [transforms.ToTensor(), transforms.Normalize(mean, std)]
+            ),
+        )
+        test_data = datasets.CIFAR10(
+            root="data",
+            train=False,
+            download=True,
+            transform=transforms.Compose(
+                [transforms.ToTensor(), transforms.Normalize(mean, std)]
+            ),
+        )
+
+        test_data_noisy = datasets.CIFAR10(
+            root="data",
+            train=False,
+            download=True,
+            transform=transforms.Compose(
+                [
+                    transforms.ToTensor(),
+                    transforms.Normalize(mean, std),
+                    AddGaussianNoise(0.0, 1.0),
+                ]
+            ),
+        )
+
+        test_data_masked = datasets.CIFAR10(
+            root="data",
+            train=False,
+            download=True,
+            transform=transforms.Compose(
+                [
+                    transforms.ToTensor(),
+                    transforms.Normalize(mean, std),
+                    AddMask(0.5, True),
+                ]
+            ),
+        )
 
     return {
         "train": torch.utils.data.DataLoader(
@@ -211,35 +239,46 @@ def test(loaders, label, device):
         print("Test Accuracy of the model on the 10000 test images: %.4f" % accuracy)
         return accuracy
 
+num_layers = np.arange(2, 5)
+num_recurrences = np.arange(3)
+num_channels = [8, 16, 24]
 
-for num_layers in range(2, 5):
-    for num_recurrence in range(3):
-        for num_channels in [8, 16, 24]:
-            seed = 0
-            random.seed(seed)
-            torch.manual_seed(seed)
-            cnn = CNN(
-                num_layers=num_layers,
-                num_recurrence=num_recurrence,
-                num_channels=num_channels,
-                dataset="cifar10",
-            ).to(device)
-            loaders = get_dataloaders()
-            loss = train(3, cnn, loaders, device)
-            test_acc = test(loaders, "test", device)
-            noisy_test_acc = test(loaders, "test_data_noisy", device)
-            masked_test_acc = test(loaders, "test_data_masked", device)
-            with open(
-                f"./CIFAR_model_num_layers={num_layers}_num_recurrence={num_recurrence}_num_channels={num_channels}.pkl",
-                "wb",
-            ) as f:
-                pickle.dump(
-                    {
-                        "cnn": cnn,
-                        "loss": loss,
-                        "test_acc": test_acc,
-                        "test_noisy_acc": noisy_test_acc,
-                        "test_masked_acc": masked_test_acc,
-                    },
-                    f,
-                )
+params = prepare_configs_loop(num_layers, num_recurrences, num_channels)
+print(len(params[0]))
+print(params)
+# for num_layers in range(2, 5):
+#     for num_recurrence in range(3):
+#         for num_channels in [8, 16, 24]:
+            # settings = [[3, 2, 16],
+
+num_layer, num_recurrence, num_channel = [params[i][args.experiment] for i in range(len(params))]
+
+if os.path.exists(f"./{dataset}_seed_{seed}_model_num_layers={num_layer}_num_recurrence={num_recurrence}_num_channels={num_channel}.pkl"):
+    print("Model already exists")
+    exit()
+
+cnn = CNN(
+    num_layers=num_layer,
+    num_recurrence=num_recurrence,
+    num_channels=num_channel,
+    dataset=dataset,
+).to(device)
+loaders = get_dataloaders(dataset)
+loss = train(3, cnn, loaders, device)
+test_acc = test(loaders, "test", device)
+noisy_test_acc = test(loaders, "test_data_noisy", device)
+masked_test_acc = test(loaders, "test_data_masked", device)
+with open(
+    f"./{dataset}_seed_{seed}_model_num_layers={num_layer}_num_recurrence={num_recurrence}_num_channels={num_channel}.pkl",
+    "wb",
+) as f:
+    pickle.dump(
+        {
+            "cnn": cnn,
+            "loss": loss,
+            "test_acc": test_acc,
+            "test_noisy_acc": noisy_test_acc,
+            "test_masked_acc": masked_test_acc,
+        },
+        f,
+    )
